requirements.txt

pandas>=2.0
requests>=2.31
simple-salesforce>=1.12
python-dateutil>=2.8

numpy>=1.24
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0
httpx
python-multipart
python-dotenv
python-jose[cryptography]
toml